Update to MediaWiki 1.44.0

+++ extensions/Echo/maintenance/removeInvalidNotification.php
@@ -19,54 +19,54 @@ require_once getenv( 'MW_INSTALL_PATH' ) !== false
  */
 class RemoveInvalidNotification extends Maintenance {
 
-	/** @var string[] */
-	protected $invalidEventType = [ 'article-linked' ];
-
 	public function __construct() {
 		parent::__construct();
 
 		$this->addDescription( "Removes invalid notifications from the database." );
+		$this->addOption( 'remove', 'Actually remove notifications instead of listing' );
 		$this->setBatchSize( 500 );
 		$this->requireExtension( 'Echo' );
 	}
 
 	public function execute() {
-		$lbFactory = DbFactory::newFromDefault();
-		if ( !$this->invalidEventType ) {
-			$this->output( "There is nothing to process\n" );
+		global $wgEchoNotifications;
 
-			return;
+		$dryRun = !$this->hasOption( 'remove' );
+
+		$lbFactory = DbFactory::newFromDefault();
+
+		$validEventTypes = array_keys( $wgEchoNotifications );
+
+		if ( $dryRun ) {
+			$this->output( "Dry run mode. Use --remove to really remove notifications.\n" );
 		}
 
 		$dbw = $lbFactory->getEchoDb( DB_PRIMARY );
 		$dbr = $lbFactory->getEchoDb( DB_REPLICA );
 
-		$batchSize = $this->getBatchSize();
-		$count = $batchSize;
+		$sqb = $dbr->newSelectQueryBuilder()
+			->select( [ 'event_id', 'event_type' ] )
+			->from( 'echo_event' )
+			->where( $dbr->expr( 'event_type', '!=', $validEventTypes ) )
+			->caller( __METHOD__ );
 
-		while ( $count == $batchSize ) {
-			$res = $dbr->newSelectQueryBuilder()
-				->select( 'event_id' )
-				->from( 'echo_event' )
-				->where( [
-					'event_type' => $this->invalidEventType,
-				] )
-				->limit( $batchSize )
-				->caller( __METHOD__ )
-				->fetchResultSet();
+		// Using echo_event_type index
+		$iterator = new BatchRowIterator( $dbr, $sqb, [ 'event_type', 'event_id' ], $this->getBatchSize() );
 
+		$countByType = [];
+		foreach ( $iterator as $batch ) {
 			$event = [];
-			$count = 0;
-			foreach ( $res as $row ) {
+			foreach ( $batch as $row ) {
 				// @phan-suppress-next-line PhanPossiblyUndeclaredVariable
 				if ( !in_array( $row->event_id, $event ) ) {
 					$event[] = $row->event_id;
+					$countByType[$row->event_type] ??= 0;
+					$countByType[$row->event_type]++;
 				}
-				$count++;
 			}
 
-			if ( $event ) {
-				$this->beginTransaction( $dbw, __METHOD__ );
+			if ( $event && !$dryRun ) {
+				$this->beginTransactionRound( __METHOD__ );
 
 				$dbw->newDeleteQueryBuilder()
 					->deleteFrom( 'echo_event' )
@@ -79,14 +79,28 @@ class RemoveInvalidNotification extends Maintenance {
 					->caller( __METHOD__ )
 					->execute();
 
-				$this->commitTransaction( $dbw, __METHOD__ );
+				$this->commitTransactionRound( __METHOD__ );
 
-				$this->output( "processing " . count( $event ) . " invalid events\n" );
-				$this->waitForReplication();
+				$this->output( "...removing " . count( $event ) . " invalid events\n" );
+			} else {
+				$this->output( "...would remove " . count( $event ) . " invalid events\n" );
 			}
 
 			// Cleanup is not necessary for
 			// 1. echo_email_batch, invalid notification is removed during the cron
+		}
+
+		if ( !$countByType ) {
+			$this->output( "Nothing to do.\n" );
+		} else {
+			if ( !$dryRun ) {
+				$this->output( "Removed " . array_sum( $countByType ) . " invalid events with types:\n" );
+			} else {
+				$this->output( "Would remove " . array_sum( $countByType ) . " invalid events with types:\n" );
+			}
+			foreach ( $countByType as $type => $removedCount ) {
+				$this->output( " - $type: $removedCount\n" );
+			}
 		}
 	}
 }
